BL-11381 Move History events to their own namespace
