+++ src/BloomExe/Publish/BloomLibrary/BloomLibraryUploadControl.cs
@@ -7,6 +7,10 @@ using System.Linq;
 using System.Windows.Forms;
 using Bloom.Book;
 using Bloom.Collection;
+using Bloom.ErrorReporter;
+using Bloom.History;
+using Bloom.ImageProcessing;
+using Bloom.MiscUI;
 using Bloom.TeamCollection;
 using Bloom.web;
 using Bloom.web.controllers;
