+++ src/BloomExe/TeamCollection/FolderTeamCollection.cs
@@ -1,23 +1,15 @@
 using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Text;
 using System.Text.RegularExpressions;
-using System.Threading;
-using System.Windows.Forms;
-using System.Xml;
-using System.Xml.Linq;
-using Bloom.Api;
 using Bloom.Collection;
 using Bloom.CollectionCreating;
 using Bloom.MiscUI;
 using Bloom.Utils;
 using Bloom.web;
 using L10NSharp;
-using Sentry;
 using SIL.Code;
 using SIL.IO;
