+++ src/BloomExe/TeamCollection/TeamCollection.cs
@@ -2,7 +2,6 @@ using Bloom.Api;
 using Bloom.MiscUI;
 using Bloom.web;
 using L10NSharp;
-using Sentry;
 using SIL.IO;
 using System;
 using System.Collections.Concurrent;
@@ -11,20 +10,18 @@ using System.ComponentModel;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Linq.Expressions;
 using System.Security.Cryptography;
 using System.Text;
-using System.Text.RegularExpressions;
 using System.Windows.Forms;
 using Bloom.Book;
 using Bloom.Collection;
+using Bloom.History;
 using Bloom.Registration;
 using Bloom.ToPalaso;
 using Bloom.Utils;
 using Bloom.web.controllers;
 using SIL.Reporting;
 using DesktopAnalytics;
-using Newtonsoft.Json;
 using SIL.Code;
 
 namespace Bloom.TeamCollection
